Move custom matchers to a single folder
