+++ spec/rails_helper.rb
@@ -13,7 +13,6 @@ require 'rspec/rails'
 require 'webmock/rspec'
 require 'support/factory_girl'
 require 'support/serializer_spec_helper'
-require 'support/xml_matchers'
 require 'support/validation_helpers'
 require 'support/model_helpers'
 require 'support/authenticated_session_helper'
