+++ spec/spec_helper.rb
@@ -3,14 +3,10 @@
 require 'fakeredis/rspec'
 require 'support/mvi/stub_mvi'
 require 'support/spec_builders'
-require 'support/schema_matchers'
+require 'support/matchers'
 require 'support/spool_helpers'
 require 'support/fixture_helpers'
 require 'support/spec_temp_files'
-require 'support/have_deep_attributes_matcher'
-require 'support/be_a_uuid'
-require 'support/impl_matchers'
-require 'support/negated_matchers'
 require 'support/sidekiq/batch'
 require 'support/stub_emis'
 require 'support/stub_evss_pciu'
