+++ spec/support/matchers.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+Dir[File.join(__dir__, 'matchers', '*.rb')].each { |file| require file }
